+++ kmer_encoder.cc
@@ -9,5 +9,5 @@ KmerEncoder::KmerEncoder()
     for (int i = 0; i < 255; i++)
 	aa_to_offset_[i] = 20;
     for (int i = 0; i < 20; i++)
-	aa_to_offset_[prot_alpha[i]] = i;
+	aa_to_offset_[(int) prot_alpha[i]] = i;
 }
